chore: move back RegisterSwaggerAPI to app
